Move error-dialog decorator from microdrop_utils to the dialogs package
microdrop_utils.decorators importing the dialogs package inverted the
utils <- application layering and dragged PySide6 into every
service/model-layer consumer of debounce/timestamped_value. The
decorator is dialog-aware, so it lives with the dialogs now; its unit
tests move to a new microdrop_application/tests home alongside it.
